Task portlet notification event uri fix
Task Portlet was originally built to live on the users private pages
in order for the notification event URIs to redirect the user to the
task portlet.
The modification will ensure that if there is no task portlet on the
users private pages the Asset renderer will look for a task portlet
in the Staff User Group private pages instead.

The following if-statement will look for a task-portlet in the Staff
User Group private pages, provided there is no task-portlet on the
users private pages.